hardcoded sorting_colums and ascending order params
These two params have a deep impact in how the visits are assign. Initially we tried to keep them as parameters
to generalize their behavior, but in practice you do not really want to change these parameters if you
want them to work correctly.

Hence, we are removing that possibility